ref(models): split up model classes into their own files
Has the __init__ file still source in all the classes as needed to minimise changes in tests and migrations
