+++ rootfs/api/models/push.py
@@ -1,0 +1,28 @@
+from django.conf import settings
+from django.db import models
+
+from api.models import UuidAuditedModel
+
+
+class Push(UuidAuditedModel):
+    """
+    Instance of a push used to trigger an application build
+    """
+    owner = models.ForeignKey(settings.AUTH_USER_MODEL)
+    app = models.ForeignKey('App')
+    sha = models.CharField(max_length=40)
+
+    fingerprint = models.CharField(max_length=255)
+    receive_user = models.CharField(max_length=255)
+    receive_repo = models.CharField(max_length=255)
+
+    ssh_connection = models.CharField(max_length=255)
+    ssh_original_command = models.CharField(max_length=255)
+
+    class Meta:
+        get_latest_by = 'created'
+        ordering = ['-created']
+        unique_together = (('app', 'uuid'),)
+
+    def __str__(self):
+        return "{0}-{1}".format(self.app.id, self.sha[:7])
